+++ miniconf_mqtt/src/lib.rs
@@ -15,10 +15,7 @@ mod tests;
 
 pub use minimq;
 
-pub use client::{
-    Activation, ChangedKey, Error, Event, Handle, Ingress, Miniconf, Publisher, QueueResult,
-    Response, ResponseQueue,
-};
+pub use client::{ChangedKey, Error, Event, Miniconf, Publisher, Service, ServiceEvent, Startup};
 
 /// Maximum path-state depth supported by `miniconf_mqtt`.
 pub const MAX_DEPTH: usize = 12;